Keep track of local addr for peer
This allow to run multiple endpoints and find the right peer
when they don't switch outgoing port

We can't use (peer_addr, local_addr) as key in the self._peers
dict as the local_addr is only know when the connection is
establisehd. It could lead to race condition when creating
connections to the same peer.