Always copy provided headers
Fixes an issue where notify message with extra headers, when sent in a
loop, ends up with a pinned CSeq number.

Since the header structure is filled in before serializing and sending
on the wire - and not copied - we end up filling in the data structure
provided in the `header` named parameter.

This effectively leaks and shares state between distinct transactions.