[peers] make Peer.close() async, actually use connection_lost
The previous closing mechanism:

- did not take into account that transports invoke connect_lost
  asynchronously
- introduced an unfortunate coupling between Peer and Connector,
  where the peer needed to know about it's connector

The new mechanism is clearer as we clearly distinguish "we want to
close the transport" and "the transport has closed". It also allows
for clients and servers to shutdown cleanly regardless of each
peer's transport.

We also stop pretending a Peer can be used as a context manager.