Queue messages in a dialog and support iteration
This API lets me work with a dialog without having to explicitly
register a callback for every message.

It also means that there's no more implicit 404 handling either, but I'm
not convinced you want to do that inside a dialog. I think that makes
sense for the dialplan, but once a dialog is established, we probably
want every incoming message to be explicitly handled, and with awareness
of the order they arrived in.