Hide `create_dialog` from the public API
Instead of creating a dialog and then sending a SUBSCRIBE or REGISTER
message on that dialog, we should couple dialog setup directly with
the SIP transaction.

It makes for both a cleaner API that's more correct (okay, well,
technically you could have always done the right thing with the old
API, but you had to know what you're doing and do the appropriate
bookkeeping yourself). If the transaction succeeds, you get a new
dialog. If it fails, the dialog is automatically cleaned up and closed
for you.

The tests and example code are both easier to follow with this change.