Dialog rework with provisonal INVITE support
Reworks Dialogs to be paired with an initial message. Removes the
private `_subscribe` and `_register` methods in favour of a new
`refresh` or `close` API.

Implement basic INVITE dialog support. This requires its own Dialog
implementation as it has its own call state information that needs to be
tracked. INVITE really is a special case.