Rework how dialogs are tracked
Its actually wrong to try and first determine the remote address
first, and then use that to work out which dialog a message belongs to
- we don't have enough information to accurately make that
determination.

All we have is sufficient information to determine which dialog and
transaction a message belongs to.

Also in the case that a new inbound message arrives and needs to be
run through the dialplan, the Contact header should be used instead of
the socket's peername for figuring out where to return traffic.

So the current design needs to be broken and globally track all
dialogs. Once we've established the dialog, we can get peer
information. Dialogs are also now tracked by the set of the address
contained in the To and From headers plus the Call-ID. This is much
closer to correct, but still a little lacking.

This isn't quite correct, and needs to shift to using the To and From
header tags, but that requires some deeper changes.

Unfortunately this completely breaks a lot of the statistics stored on
the peers, and I've removed them for the time being.

Closes #93