Rework dialplan
Current dialplan was not correct and complicated. This rework
simplify it greatly and provide an easy way to do your own routing

I think it will be a nice place to put helpers for
authentifications.

No idea if that will provide any advantages over a flat mapping
of methods / middlewares

issue #78